# apps/web/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vacation Price Tracker",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/trips": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "List Trips",
        "description": "List trips for the current user with latest snapshot prices.",
        "operationId": "list_trips_v1_trips_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TripStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_list_TripResponse__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Create Trip",
        "description": "Create a new trip for the current user.",
        "operationId": "create_trip_v1_trips_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_TripDetail_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/trips/refresh-status": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Refresh Status",
        "description": "Check refresh workflow progress.",
        "operationId": "refresh_status_v1_trips_refresh_status_get",
        "parameters": [
          {
            "name": "refresh_group_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Refresh Group Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_RefreshStatusResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/trips/{trip_id}": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Get Trip Details",
        "description": "Get trip details including prefs, notification rules, and price history.",
        "operationId": "get_trip_details_v1_trips__trip_id__get",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_TripDetailResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "trips"
        ],
        "summary": "Delete Trip",
        "description": "Delete a trip and all associated records.",
        "operationId": "delete_trip_v1_trips__trip_id__delete",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/trips/{trip_id}/status": {
      "patch": {
        "tags": [
          "trips"
        ],
        "summary": "Update Trip Status",
        "description": "Pause or resume trip tracking.",
        "operationId": "update_trip_status_v1_trips__trip_id__status_patch",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Trip Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_TripResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponse_RefreshStatusResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RefreshStatusResponse"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIResponse[RefreshStatusResponse]"
      },
      "APIResponse_TripDetailResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TripDetailResponse"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIResponse[TripDetailResponse]"
      },
      "APIResponse_TripDetail_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TripDetail"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIResponse[TripDetail]"
      },
      "APIResponse_TripResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TripResponse"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIResponse[TripResponse]"
      },
...
      "APIResponse_list_TripResponse__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/TripResponse"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "APIResponse[list[TripResponse]]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TripCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "Trip name (must be unique per user)"
          },
          "origin_airport": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "title": "Origin Airport",
            "description": "Origin airport IATA code (e.g., 'SFO')"
          },
          "destination_code": {
            "type": "string",
            "pattern": "^[A-Z]{3}$",
            "title": "Destination Code",
            "description": "Destination airport IATA code (e.g., 'MCO')"
          },
          "is_round_trip": {
            "type": "boolean",
            "title": "Is Round Trip",
            "description": "Whether this is a round trip",
            "default": true
          },
          "depart_date": {
            "type": "string",
            "format": "date",
            "title": "Depart Date",
            "description": "Departure date"
          },
          "return_date": {
            "type": "string",
            "format": "date",
            "title": "Return Date",
            "description": "Return date"
          },
          "adults": {
            "type": "integer",
            "maximum": 9.0,
            "minimum": 1.0,
            "title": "Adults",
            "description": "Number of adult travelers",
            "default": 1
          },
          "flight_prefs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FlightPrefs"
              },
              {
                "type": "null"
              }
            ],
            "description": "Flight preferences (optional)"
          },
          "hotel_prefs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HotelPrefs"
              },
              {
                "type": "null"
              }
            ],
            "description": "Hotel preferences (optional)"
          },
          "notification_prefs": {
            "$ref": "#/components/schemas/NotificationPrefs-Input",
            "description": "Notification settings (required)"
          }
        },
        "type": "object",
        "required": [
          "name",
          "origin_airport",
          "destination_code",
          "depart_date",
          "return_date",
          "notification_prefs"
        ],
        "title": "TripCreate",
        "description": "Schema for creating a new trip."
      },
...
      "TripDetailResponse": {
        "properties": {
          "trip": {
            "$ref": "#/components/schemas/TripDetail"
          },
          "price_history": {
            "items": {
              "$ref": "#/components/schemas/PriceSnapshotResponse"
            },
            "type": "array",
            "title": "Price History"
          }
        },
        "type": "object",
        "required": [
          "trip",
          "price_history"
        ],
        "title": "TripDetailResponse",
        "description": "Schema for trip detail response with price history."
      },
      "TripResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "origin_airport": {
            "type": "string",
            "title": "Origin Airport"
          },
          "destination_code": {
            "type": "string",
            "title": "Destination Code"
          },
          "depart_date": {
            "type": "string",
            "format": "date",
            "title": "Depart Date"
          },
          "return_date": {
            "type": "string",
            "format": "date",
            "title": "Return Date"
          },
          "status": {
            "$ref": "#/components/schemas/TripStatus"
          },
          "current_flight_price": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Flight Price"
          },
          "current_hotel_price": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Hotel Price"
          },
          "total_price": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Price"
          },
          "last_refreshed": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Refreshed"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "origin_airport",
          "destination_code",
          "depart_date",
          "return_date",
          "status"
        ],
        "title": "TripResponse",
        "description": "Schema for trip list response."
      },
      "TripStatus": {
        "type": "string",
        "enum": [
          "active",
          "paused",
          "error"
        ],
        "title": "TripStatus",
        "description": "Trip tracking status values."
      },
      "TripStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "active",
              "paused"
            ],
            "title": "Status",
            "description": "New status: 'active' or 'paused'"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "TripStatusUpdate",
        "description": "Schema for updating trip status (pause/resume)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}